Minor updates for diagnostic purposes

+++ src/org/usfirst/frc/team4256/robot/Drive4256.java
@@ -121,7 +121,7 @@ public class Drive4256 {
 	
 	public void lockAngle(boolean updateAngle) {
 		if(updateAngle && lockedAngle == null) {
-			lockedAngle = Robot.gyro.getCurrentAngle();
+			lockedAngle = (double)Robot.gyro.getCurrentAngle();
 		}else{
 			lockedAngle = null;
 		}
